requirements.txt
fastapi>=0.95.0
uvicorn>=0.22.0
sqlalchemy>=2.0.9
python-dotenv>=1.0.0
psycopg2-binary>=2.9.6
pydantic>=2.0.0
scikit-learn>=1.2.2
numpy>=1.24.3
pandas>=2.0.0